Category & posts delete fix

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -43,6 +43,10 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
+        if ($category->posts()->count()) {
+            return back()->withErrors(['error' => 'Cannot delete, category has posts.']);
+        }
+
         $category->delete();
 
         return redirect()->route('admin.categories.index');
