+++ app/Http/Controllers/Admin/TagController.php
@@ -43,7 +43,13 @@ class TagController extends Controller
 
     public function destroy(Tag $tag)
     {
-        $tag->delete();
+        foreach ($tag->posts as $post) {
+            $post->tags()->detach();
+        }
+
+        if (!$tag->posts()->count()) {
+            $tag->delete();
+        }
 
         return redirect()->route('admin.tags.index');
     }
